refactor: NetworkLog 데이터 클래스를 storage/model 패키지로 이동
- NetworkLog 는 도메인 모델 없는 순수 디버깅 데이터 클래스이므로 LocalEntity 와 도메인 분리가 의미 없음 → 별도 모델 없이 위치만 이동
- 결과적으로 storage 직렬화 대상이 모두 storage.model.** 한 패키지에 모임 (proguard keep 좁힘에 유리)
- 사용처 (NetworkModule, NetworkModuleForGoogle, SNUTTStorage, SNUTTStorageExt, DebugViewModel, NetworkLogPage, NetworkLogger) import 일괄 갱신
