+++ app/src/main/java/com/wafflestudio/snutt2/feature/settings/DebugViewModel.kt
@@ -1,8 +1,8 @@
 package com.wafflestudio.snutt2.feature.settings
 
 import androidx.lifecycle.ViewModel
-import com.wafflestudio.snutt2.lib.android.NetworkLog
 import com.wafflestudio.snutt2.storage.SNUTTStorage
+import com.wafflestudio.snutt2.storage.model.NetworkLog
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
