+++ app/src/main/java/com/wafflestudio/snutt2/storage/SNUTTStorageExt.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.snutt2.storage
 
-import com.wafflestudio.snutt2.lib.android.NetworkLog
+import com.wafflestudio.snutt2.storage.model.NetworkLog
 
 fun SNUTTStorage.addNetworkLog(newLog: NetworkLog) {
     networkLog.update(
